pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pavel Kozlov <mail@pkozlov.ru>"]

[tool.poetry.dependencies]
python = "^3.8.5"
django = "^3.1.2"
psycopg2 = "^2.8.6"
djangorestframework = "^3.12.1"
markdown = "^3.3.3"
django-filter = "^2.4.0"
djangorestframework-gis = "^0.16"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
